pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-trend-news"
version = "0.1.0"
description = "Daily topic-based news PDF generator with Gmail delivery."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "feedparser>=6.0.11",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
  "reportlab>=4.2.0",
  "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
trend-news = "trend_news.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
